HU16 se organizaron los endpoinnts en los controller
